Implemented AcmeCache trait
This uses async_trait to expose an interface to implement custom cache
backends. An implementation for the filesystem is still provided.

Existing code should work -- in the hyper_rustls example, filling in the
cache_dir setting compiles properly. The changes in the example file are
primarly updates to support rustls 0.20.