Moved cache error type to the cache mod
While the `AnyError` trait is generic enough to be used in all
situations its needed, we only use it in the context of the cache
currently. It seems like a better fit to be renamed and live in this
module -- and it makes the module not just contain one item.